+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/web3j/privacy/OnChainPrivacyAcceptanceTestBase.java
@@ -24,40 +24,24 @@ import org.hyperledger.besu.tests.acceptance.dsl.privacy.PrivacyNode;
 import org.hyperledger.besu.tests.acceptance.dsl.privacy.condition.ExpectValidOnChainPrivacyGroupCreated;
 import org.hyperledger.besu.tests.acceptance.dsl.privacy.transaction.CreateOnChainPrivacyGroupTransaction;
 import org.hyperledger.besu.tests.acceptance.dsl.transaction.privacy.PrivacyRequestFactory;
-import org.hyperledger.enclave.testutil.EnclaveType;
 
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.apache.tuweni.bytes.Bytes;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.web3j.protocol.besu.response.privacy.PrivateTransactionReceipt;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
 import org.web3j.tx.Contract;
 import org.web3j.utils.Base64String;
 
-@RunWith(Parameterized.class)
 public class OnChainPrivacyAcceptanceTestBase extends PrivacyAcceptanceTestBase {
-
-  protected final EnclaveType enclaveType;
-
-  public OnChainPrivacyAcceptanceTestBase(final EnclaveType enclaveType) {
-    this.enclaveType = enclaveType;
-  }
-
-  @Parameterized.Parameters(name = "{0}")
-  public static Collection<EnclaveType> enclaveTypes() {
-    return Arrays.asList(EnclaveType.values());
-  }
 
   protected String createOnChainPrivacyGroup(final PrivacyNode... members) {
     final List<String> addresses =
-        Arrays.asList(members).stream().map(m -> m.getEnclaveKey()).collect(Collectors.toList());
+        Arrays.stream(members).map(PrivacyNode::getEnclaveKey).collect(Collectors.toList());
     return createOnChainPrivacyGroup(members[0].getEnclaveKey(), addresses, members);
   }
